chore 📦: Add dependencies, classifiers, and scripts to pyproject.toml
This commit updates the `pyproject.toml` file to include new dependencies for `typer`, `rich`, and `faster-whisper`. It also adds a new package entry in the `[tool.setuptools]` section. Additionally, it introduces a new script in the `[tool.pdm.scripts]` section to prepare the ROCm shared library into the installed ctranslate2 package and set the RPATH. These changes enhance the project's functionality by providing additional dependencies and scripts for better integration with ROCm environments.